Prevent collisions by using $apply over $digest
I was having a problem where the new elements added by an ng-repeat were
showing, but their bindings were applied a bit later, meaning that 
they'd show as "{{ something }}" for a moment first. I eventually traced
it back to this click event handler, and found that doing an $apply
instead of a $digest solved the problem.